barrett15_74: add workaround for MacOS bug
This fixes failures in barrett15_74 on MacOS 14 with Radeon Pro 560.

Keep 2 extra bits for a.d4 before loop just like it's done later inside
the loop. It may not be necessary mathematically, but it prevents an
optimization bug that turns a.d4 into a random number on MacOS.

Recognize "Radeon Pro 560" that exhibits the issue as GCN4.