Unify makefiles for Linux and MacOS
Don't pass BITS to the compiler unless bitness is overridden.

Don't try to support generic Windows. MSYS has the install command, MSVC
doesn't use the Makefile.

Use the install command unconditionally. Make mfakto.ini writable when
installed.

Set the optimization flags based on whether the compiler is clang or
gcc, not based on the OS.

Compile C++ files with the C++ compiler.

Properly mark targets as phony, the syntax was incorrect. The "src/all"
file is not created anymore.

Don't override CFLAGS and LDFLAGS on MacOS in CI.